flyttet mye av koden over i api.ts,
