+++ frontend/altinn-support-dashboard.client/src/hooks/hooks.ts
@@ -20,7 +20,7 @@ import {
   fetchOrganizations,
   fetchPersonalContacts,
   fetchRoles,
-  fetchRoleTypes,
+  fetchRolesForOrg,
   fetchSubunits,
 } from "../utils/api";
 import { OfficialContact } from "../components/Dashboard/models/mainContentTypes";
@@ -140,23 +140,7 @@ export function UseManualRoleSearch(rollehaver: string, rollegiver: string) {
 
   return useQuery ({
     queryKey: ["manualroles", baseUrl, rollehaver, rollegiver],
-    queryFn: async () => {
-      if (!rollehaver || !rollegiver) return [];
-      const res = await fetch( `/api/TT02/serviceowner/${rollehaver}/roles/${rollegiver}`);
-      if (!res.ok) { const errorText = await res.text(); throw new Error(errorText || "Error fetching roles");}
-      const data = await res.json();
-      let rolesArray: Role[] = [];
-      if (Array.isArray(data)) {
-        rolesArray = data;
-      } else if (data && data._embedded) {
-        const embeddedKeys = Object.keys(data._embedded);
-        if (embeddedKeys.length > 0) {
-          const firstKey = embeddedKeys[0];
-          rolesArray = data._embedded[firstKey];
-        }
-      }
-      return rolesArray;
-    },
+    queryFn: async () => fetchRolesForOrg(environment, rollehaver, rollegiver),
     enabled: !!rollehaver && !!rollegiver,
   });
 };
